re-organize the io sub-package
