+++ lightkurve/io/__init__.py
@@ -1,2 +1,3 @@
-from .core import *
+from .detect import *
+from .official_products import *
 from .read import *
