+++ lightkurve/io/read.py
@@ -4,12 +4,14 @@ import logging
 from astropy.io import fits
 from astropy.utils import deprecated
 
+from .detect import detect_filetype
 from ..lightcurve import KeplerLightCurve, TessLightCurve
 from ..utils import validate_method, LightkurveWarning, LightkurveDeprecationWarning
 
 log = logging.getLogger(__name__)
 
-__all__ = ['open', 'read', 'detect_filetype']
+
+__all__ = ['open', 'read']
 
 
 @deprecated("2.0", alternative="read()", warning_type=LightkurveDeprecationWarning)
@@ -83,83 +85,3 @@ def read(path_or_url, **kwargs):
         # if these keywords don't exist, raise `ValueError`
         raise ValueError("Not recognized as a Kepler or TESS data product: "
                          "{}".format(path_or_url))
-
-
-def detect_filetype(hdulist):
-    """Returns Kepler and TESS file types given a FITS object.
-
-    This function will detect the file type by looking at both the TELESCOP and
-    CREATOR keywords in the first extension of the FITS header. If the file is
-    recognized as a Kepler or TESS data product, one of the following strings
-    will be returned:
-
-        * `'KeplerTargetPixelFile'`
-        * `'TessTargetPixelFile'`
-        * `'KeplerLightCurve'`
-        * `'TessLightCurve'`
-
-    In addition, community-provided data products such as K2SFF are supported.
-
-    If the data product cannot be detected, `None` will be returned.
-
-    Parameters
-    ----------
-    hdulist : astropy.io.fits.HDUList object
-        A FITS file.
-
-    Returns
-    -------
-    filetype : str or None
-        A string describing the detected filetype. If the filetype is not
-        recognized, `None` will be returned.
-    """
-    # Is it a K2SFF file?
-    try:
-        # There are no metadata keywords identifying K2SFF FITS files,
-        # so we go by structure.
-        if hdulist[1].header['EXTNAME'] == "BESTAPER" and hdulist[1].header["TTYPE4"] == "ARCLENGTH":
-            return "K2SFF"
-    except Exception:
-        pass
-
-    # Is it an EVEREST file?
-    try:
-        if "EVEREST" in str(hdulist[0].header['COMMENT']):
-            return "EVEREST"
-    except Exception:
-        pass
-
-    # Is it an official data product?
-    header = hdulist[0].header
-    try:
-        # use `telescop` keyword to determine mission
-        # and `creator` to determine tpf or lc
-        if 'TELESCOP' in header.keys():
-            telescop = header['telescop'].lower()
-        else:
-            # Some old custom TESS data did not define the `TELESCOP` card
-            telescop = header['mission'].lower()
-        creator = header['creator'].lower()
-        origin = header['origin'].lower()
-        if telescop == 'kepler':
-            # Kepler TPFs will contain "TargetPixelExporterPipelineModule"
-            if 'targetpixel' in creator:
-                return 'KeplerTargetPixelFile'
-            # Kepler LCFs will contain "FluxExporter2PipelineModule"
-            elif ('fluxexporter' in creator or 'lightcurve' in creator
-                or 'lightcurve' in creator):
-                return 'KeplerLightCurve'
-        elif telescop == 'tess':
-            # TESS TPFs will contain "TargetPixelExporterPipelineModule"
-            if 'targetpixel' in creator:
-                return 'TessTargetPixelFile'
-            # TESS LCFs will contain "LightCurveExporterPipelineModule"
-            elif 'lightcurve' in creator:
-                return 'TessLightCurve'
-            # Early versions of TESScut did not set a good CREATOR keyword
-            elif 'stsci' in origin:
-                return 'TessTargetPixelFile'
-    # If the TELESCOP or CREATOR keywords don't exist we expect a KeyError;
-    # if one of them is Undefined we expect `.lower()` to yield an AttributeError.
-    except (KeyError, AttributeError):
-        return None
